requirements.txt
colorama==0.4.6
dnspython==2.7.0
iniconfig==2.1.0
mongoengine==0.29.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pymongo==4.14.1
pytest==8.4.1
python-dotenv==1.2.1
pika==1.3.2

